% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT.cg.R
\name{PLOT.cg}
\alias{PLOT.cg}
\title{Fatigue Crack Growth in Reliability plots}
\usage{
PLOT.cg(x)
}
\arguments{
\item{x}{cracks.growth object.}
}
\value{
\describe{ Return the following values:
\item{\code{plot.data}}{Exploratory chart.}
\item{\code{plot.pred}}{Plot for fatigue lifetimes estimates and predictions.}
\item{\code{plot.F}}{Plot for the empirical distribution and lifetimes
distribution estimates of fatigue lifetimes.}
\item{\code{plot.resid}}{Residuals chart.}
}
}
\description{
It provides graphical outputs composed of the trends corresponding to the
crack length growth due to mechanical fatigue, the crack length estimates by the models,
crack length predictions, and lifetime distribution estimates.
}
\details{
Specifically, the following graphs are provided: exploratory dataset graph,
plot with the crack length estimates and predictions, residuals graph, empirical
and estimated lifetime distribution plot obtained by SEP-lme_bkde, SEP-lme_kde or
PB-nlme methods.
}
\examples{

## Using the Alea.A dataset
data(Alea.A)
x <- Alea.A
## Critical crack length
aF <- 1.6
## Censoring time
T_c <- 0.12
## cracks.growth function applied to Alea.A data
cg <- cracks.growth(x, aF, T_c,
                   method = c("SEP-lme_bkde", "SEP-lme_kde", "PB-nlme"),
                   nBKDE = 1000, nKDE = 1000, nMC = 5000)
## PLOT.cg applied to cg object.
PLOT <- PLOT.cg(cg)
names(PLOT)
## [1]  "plot.data"  "plot.pred"  "plot.F"     "plot.resid"
## Exploratory chart for the Alea.A dataset
PLOT$plot.data(main = "Plot:  crack growth", xlab = "million cycles",
               ylab = "cracks(inches)",  cex.lab=1.8,
               cex.main = 2)
text(0.02, aF + 0.05, "Failure", cex = 1.8)
text(0.095, 0.95, "Censoring time->", cex = 1.5)
## Plot for fatigue lifetimes estimates and predictions.
PLOT$plot.pred(xlab = "million cycles", ylab = "cracks(inches)",
               main = "Plot: crack growth, estimation and prediction\n failure times (red)",
               cex.lab = 1.8, cex.main = 1.5)
text(0.02,aF+0.05, "Failure", cex = 1.8)
text(0.085,0.95, "Censoring time->", cex = 1.5)
## Plot for the empirical distribution and lifetimes distribution estimates
## of  fatigue lifetimes
PLOT$plot.F(main = "Plot: distributions of failure times",
            xlab = "million cycles", ylab = "probability",
            cex.lab = 1.7, cex.main=2)
text(0.14, 0.1, "<-Censoring time", cex = 1.5)
legend("topleft", c("Empirical", "Estimated"), col = c("blue","black"),
       pch=c(20,20), cex=1.5, bty="n")
## Residuals chart.
PLOT$plot.resid(main = "Plot: residual", xlab = "fitted", ylab = "residuals",
                cex = 1.5, col = "blue", cex.lab = 1.7, cex.main = 2)
}
\references{
Meeker, W., Escobar, L. (1998) Statistical Methods for Reliability Data. John Wiley & Sons,
Inc. New York.

Pinheiro JC., Bates DM. (2000) Mixed-effects models in S ans S-plus. Statistics and Computing.
Springer-Verlang. New York.

Paris, P.C. and Erdogan, F. (1963) A critical analysis of crack propagation laws. J. Basic Eng.,
85, 528.
}
\author{
Antonio Meneses antoniomenesesfreire@hotmail.com, Salvador Naya salva@udc.es,
Javier Tarrio-Saavedra jtarrio@udc.es, Ignacio Lopez-Ullibarri ilu@udc.es
}
